requirements.txt
# Core ML / YOLO
ultralytics>=8.3.0
torch>=2.0.0
torchvision>=0.15.0
onnx>=1.16.0
onnxruntime>=1.18.0

# Data handling & versioning
dvc>=3.0
dvc-gdrive>=3.0.0

# Web UI & API
streamlit>=1.38.0
fastapi>=0.115.0
uvicorn>=0.30.0
python-multipart>=0.0.9

# Image processing & augmentation
opencv-python>=4.8.0
pillow>=10.0.0
albumentations>=1.4.0

# Experiment tracking & logging
wandb>=0.18.0

# Utilities
pyyaml>=6.0
tqdm>=4.66.0
pandas>=2.0.0
numpy>=1.24.0

# Optional visualization
folium>=0.17.0
matplotlib>=3.7.0
